The distributions lab + inference lab is updated
The distributions lab appears that it should work with the current sequence of distributions lessons and somewhat with the CTL lesson.  Questions 2 & 7 are meant to attempt after completing the CTL material since those questions are mentioned explicitly in that content.

The lab for Inference does make use of Pandas so I may change this out tomorrow / update it but for now, this is what I can do given the timeframe.